requirements.txt
Django==5.2.8
djangorestframework==3.14.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
dj-database-url==2.1.0
